+++ src/features/login/components/LoginTitle.tsx
@@ -1,21 +1,18 @@
-import { SafeAreaView } from "react-native-safe-area-context";
 import styled from "styled-components/native";
 import { Image as Img } from "expo-image";
 import responsiveToPx from "@/src/utils/responsiveToPx";
 
-export const SafeView = styled(SafeAreaView)`
-  flex: 1;
-  background-color: ${({ theme }) => theme.colors.whiteBlue};
-`;
+export default function LoginTitle() {
+  return (
+    <StyledBox>
+      <Image source={require("@/assets/images/logo.svg")} contentFit="contain" />
+      <MelissaText>Melissa</MelissaText>
+      <TitleText>하루의 끝, 멜리사</TitleText>
+    </StyledBox>
+  );
+}
 
-export const ContentBox = styled.View`
-  flex: 1;
-  background-color: ${({ theme }) => theme.colors.whiteBlue};
-  padding: ${responsiveToPx("80px")} 0px;
-  justify-content: space-between;
-`;
-
-export const TextBox = styled.View`
+const StyledBox = styled.View`
   align-self: center;
   flex: 1;
   justify-content: center;
@@ -24,26 +21,20 @@ export const TextBox = styled.View`
   gap: ${responsiveToPx("8px")};
 `;
 
-export const Image = styled(Img)`
+const Image = styled(Img)`
   width: ${responsiveToPx("88px")};
   height: ${responsiveToPx("70px")};
   justify-content: center;
   align-items: center;
 `;
 
-export const TitleText = styled.Text`
+const TitleText = styled.Text`
   font-family: ${({ theme }) => theme.fontFamily.nsRegular};
   font-size: ${({ theme }) => theme.fontSize.md};
   color: ${({ theme }) => theme.colors.deepGreen};
 `;
 
-export const MelissaText = styled(TitleText)`
+const MelissaText = styled(TitleText)`
   font-family: ${({ theme }) => theme.fontFamily.poetsenOne};
   font-size: ${({ theme }) => theme.fontSize.xxxxl};
-`;
-
-export const ButtonBox = styled.View`
-  justify-content: center;
-  align-items: center;
-  gap: ${({ theme }) => theme.gap.lg};
 `;
