+++ src/libs/mmkv.ts
@@ -6,10 +6,6 @@ const setStorageValue = (key: string, value: string): void => {
   storage.set(key, value);
 };
 
-const getStorageValue = (key: string): string | null => {
-  return storage.getString(key) || null;
-};
-
 const removeStorageValue = (key: string): void => {
   storage.delete(key);
 };
